+++ apps/src/componentLibrary/common/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.2](https://github.com/code-dot-org/code-dot-org/pull/62862)
+
+- updated color variables to use `primitiveColors.css`
+
+## [0.5.1](https://github.com/code-dot-org/code-dot-org/pull/62743)
+
+- updated comments for `colors.css` and `primitiveColors.css` DSCO Variables colors
+
 ## [0.5.0](https://github.com/code-dot-org/code-dot-org/pull/62334)
 
 - created `hooks` folder